+++ tests/files/table-comments/input.lua
@@ -1,0 +1,5 @@
+local x =  {
+    foo, -- comment
+    bar, -- test
+    baz
+}